writing/requirements.txt
volcengine-python-sdk[ark]>=1.0.0
questionary>=2.0.1
rich>=13.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
flask>=2.0.0
